+++ apps/website/src/routes/website/(dashboard)/editor/new/+page.svelte
